Revert "Merge branch 'Driver-update' into Test"
This reverts commit 9c1fa2b3f0034b28bef91b1fc914c05f3247677a, reversing
changes made to 5102ef59017610f88f3d372a953355a5e7df4f50.
