+++ Sensor/Encryption.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"encoding/base64"
+)
+
+func toBase64(str string) string {
+	return base64.StdEncoding.EncodeToString([]byte(str))
+}
