+++ Sensor/go.mod
@@ -2,4 +2,4 @@ module arduinobrain
 
 go 1.16
 
-require tinygo.org/x/drivers v0.23.0
+require github.com/tuxago/go v0.0.0-20221013123331-992ac8b4e6ce
