--- Sensor2/go.mod
@@ -1,3 +1,0 @@
-module SmartSignage
-
-go 1.16
